fix: package refactor and remove duplicated DataProvider
