+++ src/main/kotlin/it/pagopa/ecommerce/helpdesk/utils/v2/SearchParamDecoderV2.kt
@@ -1,5 +1,6 @@
-package it.pagopa.ecommerce.helpdesk.utils
+package it.pagopa.ecommerce.helpdesk.utils.v2
 
+import it.pagopa.ecommerce.helpdesk.utils.ConfidentialMailUtils
 import it.pagopa.generated.ecommerce.helpdesk.v2.model.EcommerceSearchTransactionRequestDto
 import it.pagopa.generated.ecommerce.helpdesk.v2.model.SearchTransactionRequestEmailDto
 import reactor.core.publisher.Mono
