+++ src/test/kotlin/it/pagopa/ecommerce/helpdesk/dataproviders/mongo/v1/EcommerceForTransactionV2DataProviderTest.kt
@@ -31,7 +31,7 @@ import it.pagopa.ecommerce.helpdesk.dataproviders.TransactionsEventStoreReposito
 import it.pagopa.ecommerce.helpdesk.dataproviders.TransactionsViewRepository
 import it.pagopa.ecommerce.helpdesk.dataproviders.v1.mongo.EcommerceTransactionDataProvider
 import it.pagopa.ecommerce.helpdesk.utils.ConfidentialMailUtils
-import it.pagopa.ecommerce.helpdesk.utils.SearchParamDecoder
+import it.pagopa.ecommerce.helpdesk.utils.v1.SearchParamDecoder
 import it.pagopa.ecommerce.helpdesk.utils.v1.getGatewayAuthorizationData
 import it.pagopa.generated.ecommerce.helpdesk.model.*
 import java.time.ZonedDateTime
